.requirements.txt
# Streamlit
streamlit==1.27.0

# Pandas
pandas==2.1.1

# Plotly et Matplotlib
plotly==5.17.0
matplotlib==3.8.1
seaborn==0.12.3

# Google Cloud BigQuery
google-cloud-bigquery==3.11.0
db-dtypes

# dates et fuseaux horaires
python-dateutil==2.8.2

# DBT API ou scripts Python
requests==2.31.0

